requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
google-generativeai>=0.8.0
pdf2image>=1.17.0
Pillow>=10.0.0
